+++ app/controllers/challah/test/restrictions_controller.rb
@@ -1,6 +1,7 @@
 # @private
 # This controller is only used for testing purposes, it does not actually get used outside of test.
 class Challah::Test::RestrictionsController < ApplicationController
+  signin_required :only => [ :blah ]
   before_filter :signin_required, :only => [ :edit ]
   restrict_to_permission :special, :only => [ :new ]
   restrict_to_authenticated :only => [ :show ]
@@ -20,6 +21,10 @@ class Challah::Test::RestrictionsController < ApplicationController
   end
 
   def edit
+    head :ok
+  end
+
+  def blah
     head :ok
   end
 end